Comments and cleaned up some other auxiliary things
