+++ node/forest_libp2p/src/rpc/mod.rs
@@ -3,16 +3,15 @@
 
 mod behaviour;
 mod codec;
+mod error;
 mod handler;
 mod protocol;
 
-pub use behaviour::*;
-pub use codec::*;
-pub use handler::*;
-pub use protocol::*;
-
-use forest_encoding::error::Error as EncodingError;
-use std::fmt;
+pub use self::behaviour::{RPCMessage, RPC};
+pub use self::codec::{InboundCodec, OutboundCodec};
+pub use self::error::RPCError;
+pub use self::handler::{RPCHandler, RESPONSE_TIMEOUT};
+pub use self::protocol::{RPCRequest, RPCResponse};
 
 pub type RequestId = usize;
 
@@ -48,38 +47,5 @@ impl std::fmt::Display for RPCEvent {
             RPCEvent::Response(id, _) => write!(f, "RPC Response(id: {:?})", id),
             RPCEvent::Error(_, err) => write!(f, "RPC Error(error: {:?})", err),
         }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum RPCError {
-    Codec(String),
-    Custom(String),
-}
-
-impl From<std::io::Error> for RPCError {
-    fn from(err: std::io::Error) -> Self {
-        Self::Custom(err.to_string())
-    }
-}
-
-impl From<EncodingError> for RPCError {
-    fn from(err: EncodingError) -> Self {
-        Self::Codec(err.to_string())
-    }
-}
-
-impl fmt::Display for RPCError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            RPCError::Codec(err) => write!(f, "Codec Error: {}", err),
-            RPCError::Custom(err) => write!(f, "{}", err),
-        }
-    }
-}
-
-impl std::error::Error for RPCError {
-    fn description(&self) -> &str {
-        "Libp2p RPC Error"
     }
 }
